Union de la tabla Book con autor y tag para relacionarlas y mantener la integridad de los datos. Nuevas querys para poder buscar por tag y autor

+++ app/src/main/java/com/echavez/bookdex/entities/Tag.kt
@@ -9,7 +9,7 @@ data class Tag(
 
         @PrimaryKey(autoGenerate = true)
         @ColumnInfo(name = "id")
-        var id: Long?= 0,
+        var id: Int?= 0,
 
         @ColumnInfo(name = "tag")
         var tag: String
